Refactor, Prune dead code
